fix(macos): regression on is_document_edited API
Ref 33fdeab6291d4aef8ea9facb58fe9583f6c1aaf3

+++ src/platform_impl/macos/window.rs
@@ -1389,7 +1389,7 @@ impl WindowExtMacOS for UnownedWindow {
 
   #[inline]
   fn set_is_document_edited(&self, edited: bool) {
-    unsafe { self.ns_window.setDocumentEdited_(edited as i8) }
+    unsafe { self.ns_window.setDocumentEdited_(edited) }
   }
 
   #[inline]
